🚚 rename: views/main -> views/mainPage 폴더명 수정
