+++ src/js/views/mainPage/contentDetailList.js
