+++ src/js/views/mainPage/contentInfo.js
